move group and address extraction to separate file
